♻️ Rework data splitting script to make it clearer
